@@ -11,7 +11,7 @@
   "author": "Tlon",
   "license": "MIT",
   "engines": {
-    "node": "16.14.0"
+    "node": "16.18.0"
   },
   "proxy": "http://localhost:4000",
   "dependencies": {
