@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 
 import Bridge from './Bridge';
 
-if (process.env.NODE_ENV !== 'production') {
-  require('@welldone-software/why-did-you-render')(React);
-}
+// if (process.env.NODE_ENV !== 'production') {
+//   require('@welldone-software/why-did-you-render')(React);
+// }
 
 ReactDOM.render(<Bridge />, document.getElementById('root'));
