@@ -1,9 +1,4 @@
-import kg from 'urbit-key-generation';
+import { expose } from "comlink";
+import { generateWallet } from './generateWalletWorker';
 
-export async function generateWallet(data: string): Promise<UrbitWallet> {
-  // Process the data without stalling the UI
-  const config = JSON.parse(data);
-  const wallet = await kg.generateWallet(config);
-
-  return wallet;
-}
+expose(generateWallet);
